streamline setup (#251)
few tweaks to make it easier to get up and running more easily:
- check whether your python v3 executable is "python" or "python3" and run that one
- load-balancer-contracts.sh only runs if folder doesn't already exist (this -allows it to be run redundantly)
- add npm run load-contracts to npm run sim which runs the above script, allowing people to skip having to run that manually
- add "export" statement to .env.sample to allow it to persist
- add instructions for setting up .env in the readme

+++ scripts/load-balancer-contracts.sh
@@ -2,24 +2,26 @@
 # Directly cloned from the script in elf frontend
 rm -rf balancer-core-v2
 
-echo "Downloading contracts..."
 # link/clone and build contracts
 if [ ! -z "$1" ] && [ $1="local" ]; then
     ln -sf ../../balancer-core-v2 .
-else
+elif [ ! -d "./contracts/balancer-core-v2" ]; then
+    echo "Downloading balancer contracts..."
     git clone https://github.com/balancer-labs/balancer-core-v2.git
     cd balancer-core-v2 \
     && git checkout f153c38c5ee8911680363eaf52aad0d691896a75 \
     && cd ..
+
+    # blow away old-contracts
+    rm -rf contracts/balancer-core-v2
+
+    echo "Copying latest contracts..."
+    mv balancer-core-v2/contracts contracts/balancer-core-v2
+
+    echo "Removing unused balancer code"
+    rm -rf balancer-core-v2
+
+    echo "Done!"
+else
+    echo "Assuming you have correctly cloned the balancer-core-v2 repo"
 fi
-
-# blow away old-contracts
-rm -rf contracts/balancer-core-v2
-
-echo "Copying latest contracts..."
-mv balancer-core-v2/contracts contracts/balancer-core-v2
-
-echo "Removing unused balancer code"
-rm -rf balancer-core-v2
-
-echo "Done!"
